refactor: Moved all dispenser behaviors including the one for equipping accessories to a designated class.
