+++ src/main/java/com/gildedgames/aether/common/item/accessories/AccessoryItem.java
@@ -1,73 +1,24 @@
 package com.gildedgames.aether.common.item.accessories;
 
 import com.gildedgames.aether.client.registry.AetherSoundEvents;
+import com.gildedgames.aether.common.registry.AetherDispenseBehaviors;
 import net.minecraft.world.level.block.DispenserBlock;
-import net.minecraft.core.dispenser.DefaultDispenseItemBehavior;
-import net.minecraft.core.BlockSource;
-import net.minecraft.core.dispenser.DispenseItemBehavior;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.Enchantments;
 import net.minecraft.world.item.Vanishable;
-import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.entity.EntitySelector;
-import net.minecraft.world.phys.AABB;
-import net.minecraft.core.BlockPos;
-import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.SlotContext;
 import top.theillusivec4.curios.api.type.capability.ICurio;
 import top.theillusivec4.curios.api.type.capability.ICurioItem;
-import top.theillusivec4.curios.api.type.inventory.ICurioStacksHandler;
-import top.theillusivec4.curios.api.type.inventory.IDynamicStackHandler;
-import top.theillusivec4.curios.api.type.util.ICuriosHelper;
 
 import javax.annotation.Nonnull;
-import java.util.List;
-import java.util.Map;
 
 public class AccessoryItem extends Item implements ICurioItem, Vanishable
 {
-    public static final DispenseItemBehavior DISPENSE_ITEM_BEHAVIOR = new DefaultDispenseItemBehavior() {
-        protected ItemStack execute(BlockSource p_82487_1_, ItemStack p_82487_2_) {
-            return AccessoryItem.dispenseAccessory(p_82487_1_, p_82487_2_) ? p_82487_2_ : super.execute(p_82487_1_, p_82487_2_);
-        }
-    }; //TODO: Move this to a registry class for these behaviors.
-
-    public static boolean dispenseAccessory(BlockSource blockSource, ItemStack stack) {
-        BlockPos blockpos = blockSource.getPos().relative(blockSource.getBlockState().getValue(DispenserBlock.FACING));
-        List<LivingEntity> list = blockSource.getLevel().getEntitiesOfClass(LivingEntity.class, new AABB(blockpos), EntitySelector.NO_SPECTATORS.and(new EntitySelector.MobCanWearArmorEntitySelector(stack)));
-        if (list.isEmpty()) {
-            return false;
-        } else {
-            LivingEntity livingentity = list.get(0);
-            ItemStack itemStack = stack.split(1);
-            ICuriosHelper curiosHelper = CuriosApi.getCuriosHelper();
-            curiosHelper.getCurio(itemStack).ifPresent(curio -> curiosHelper.getCuriosHandler(livingentity).ifPresent(handler -> {
-                Map<String, ICurioStacksHandler> curios = handler.getCurios();
-                for (Map.Entry<String, ICurioStacksHandler> entry : curios.entrySet()) {
-                    IDynamicStackHandler stackHandler = entry.getValue().getStacks();
-                    for (int i = 0; i < stackHandler.getSlots(); i++) {
-                        String id = entry.getKey();
-                        SlotContext slotContext = new SlotContext(id, livingentity, i, true, true);
-                        if (curiosHelper.isStackValid(slotContext, itemStack) && curio.canEquip(id, livingentity) && curio.canEquipFromUse(slotContext)) {
-                            ItemStack present = stackHandler.getStackInSlot(i);
-                            if (present.isEmpty()) {
-                                stackHandler.setStackInSlot(i, itemStack.copy());
-                                int count = itemStack.getCount();
-                                itemStack.shrink(count);
-                            }
-                        }
-                    }
-                }
-            }));
-            return true;
-        }
-    }
-
     public AccessoryItem(Properties properties) {
         super(properties);
-        DispenserBlock.registerBehavior(this, DISPENSE_ITEM_BEHAVIOR);
+        DispenserBlock.registerBehavior(this, AetherDispenseBehaviors.DISPENSE_ACCESSORY_BEHAVIOR);
     }
 
     @Override
